Runtime でコンパイルが通るように UnityEditor を ifdef で囲む
